+++ test-app/spec/integration/openapi3_spec.rb
@@ -5,11 +5,11 @@ require 'rswag/specs/swagger_formatter'
 # Specifically here, we look at OpenApi 3 as documented at
 # https://swagger.io/docs/specification/about/
 
-RSpec.describe 'Generated OpenApi', type: :request, swagger_doc: 'v3/openapi.json' do
+RSpec.describe 'Generated OpenApi', type: :request, openapi_spec: 'v3/openapi.json' do
   before do |example|
     output = double('output').as_null_object
-    swagger_root = File.expand_path('tmp/swagger', __dir__)
-    config = double('config', swagger_root: swagger_root, get_swagger_doc: swagger_doc )
+    openapi_root = File.expand_path('tmp/swagger', __dir__)
+    config = double('config', openapi_root: openapi_root, get_openapi_spec: openapi_spec)
     formatter = Rswag::Specs::SwaggerFormatter.new(output, config)
 
     example_group = OpenStruct.new(group: OpenStruct.new(metadata: example.metadata))
@@ -17,7 +17,7 @@ RSpec.describe 'Generated OpenApi', type: :request, swagger_doc: 'v3/openapi.jso
   end
 
   # Framework definition, to be overridden for contexts
-  let(:swagger_doc) do
+  let(:openapi_spec) do
     { # That which would be defined in swagger_helper.rb
       openapi: api_openapi,
       info: {},
@@ -42,7 +42,7 @@ RSpec.describe 'Generated OpenApi', type: :request, swagger_doc: 'v3/openapi.jso
           run_test!
 
           it 'lists server' do
-            tree = swagger_doc.dig(:servers)
+            tree = openapi_spec.dig(:servers)
             expect(tree).to eq([
               { url: "https://api.example.com/foo" }
             ])
@@ -55,7 +55,7 @@ RSpec.describe 'Generated OpenApi', type: :request, swagger_doc: 'v3/openapi.jso
             ]}
 
             it 'lists servers' do
-              tree = swagger_doc.dig(:servers)
+              tree = openapi_spec.dig(:servers)
               expect(tree).to eq([
                 { url: "https://api.example.com/foo" },
                 { url: "http://api.example.com/foo" }
@@ -74,7 +74,7 @@ RSpec.describe 'Generated OpenApi', type: :request, swagger_doc: 'v3/openapi.jso
             }]}
 
             it 'lists server and variables' do
-              tree = swagger_doc.dig(:servers)
+              tree = openapi_spec.dig(:servers)
               expect(tree).to eq([{
                 url: "https://{defaultHost}/foo",
                 variables: {
@@ -102,7 +102,7 @@ RSpec.describe 'Generated OpenApi', type: :request, swagger_doc: 'v3/openapi.jso
 
           it 'declares output as application/json' do
             pending "Not yet implemented?"
-            tree = swagger_doc.dig(:paths, "/stubs", :get, :responses, '200', :content)
+            tree = openapi_spec.dig(:paths, "/stubs", :get, :responses, '200', :content)
             expect(tree).to have_key('application/json')
           end
         end
@@ -129,13 +129,13 @@ RSpec.describe 'Generated OpenApi', type: :request, swagger_doc: 'v3/openapi.jso
             run_test!
 
             it 'declares parameter in path' do
-              tree = swagger_doc.dig(:paths, "/stubs/{a_param}", :get, :parameters)
+              tree = openapi_spec.dig(:paths, "/stubs/{a_param}", :get, :parameters)
               expect(tree.first[:name]).to eq('a_param')
               expect(tree.first[:in]).to eq(:path)
             end
 
             it 'declares path parameters as required' do
-              tree = swagger_doc.dig(:paths, "/stubs/{a_param}", :get, :parameters)
+              tree = openapi_spec.dig(:paths, "/stubs/{a_param}", :get, :parameters)
               expect(tree.first[:required]).to eq(true)
             end
           end
@@ -159,7 +159,7 @@ RSpec.describe 'Generated OpenApi', type: :request, swagger_doc: 'v3/openapi.jso
             run_test!
 
             it 'declares parameter in query string' do
-              tree = swagger_doc.dig(:paths, "/stubs", :get, :parameters)
+              tree = openapi_spec.dig(:paths, "/stubs", :get, :parameters)
               expect(tree.first[:name]).to eq('a_param')
               expect(tree.first[:in]).to eq(:query)
             end
@@ -195,7 +195,7 @@ RSpec.describe 'Generated OpenApi', type: :request, swagger_doc: 'v3/openapi.jso
 
           it 'declares requestBody is required' do
             pending "This output is massaged in SwaggerFormatter#stop, and isn't quite ready here to assert"
-            tree = swagger_doc.dig(:paths, "/stubs", :post, :requestBody)
+            tree = openapi_spec.dig(:paths, "/stubs", :post, :requestBody)
             expect(tree[:required]).to eq(true)
           end
         end